Split content definition event handling out of OmnichannelContentTypeProvider
OmnichannelContentTypeProvider resolved IContentDefinitionManager while
also being registered as an IContentDefinitionEventHandler, which created
a circular dependency because resolving the manager resolves its event
handlers. Move the event handling into a dedicated
OmnichannelContentTypeCacheInvalidator that only removes the per-tenant
memory cache entry when a content definition changes, so the provider is
left with just the cached async query methods.

The provider recomputes fresh membership sets on the next read after the
cache is invalidated, so the copy-on-write logic on OmnichannelContentTypeSet
is no longer needed and has been removed.
