+++ src/CrestApps.Docs/docs/changelog/3.0.0.md
@@ -26,7 +26,7 @@ This page highlights the changes in the `3.0.0` line.
   | `IReadOnlyCollection<string> GetSubjectContentTypes()` | `ValueTask<IReadOnlyCollection<string>> GetSubjectContentTypesAsync()` |
   | `IReadOnlyCollection<string> GetContactContentType()` | `ValueTask<IReadOnlyCollection<string>> GetContactContentTypesAsync()` |
 
-  The provider now caches its membership sets in a per-tenant `IMemoryCache`, so the initialization is handled internally and callers should simply `await` the query methods.
+  The provider now caches its membership sets in a per-tenant `IMemoryCache`, so the initialization is handled internally and callers should simply `await` the query methods. `OmnichannelContentTypeProvider` no longer implements `IContentDefinitionEventHandler`; cache invalidation on content definition changes is handled by a separate internal handler.
 
 -----
